Go back to 1s wait for ready
The old way to detect whether the bot is ready
was to wait for a reply to our "ping" package.
Sadly, with a bigger buddylist, this reply came
in the middle of the initial buddylist update,
so it was unsuited for detecting readiness.
We're now treating the client ot be ready if
no package was received for 1s.